tests/submission/W3C/domhtml/HTMLStyleElement03.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<style media="screen" type="text/css"></style>
<title>NIST DOM HTML Test - STYLE</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-style-element.html#dom-style-type" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>Hello, World.</p>



<script>
var domtest = async_test("The type attribute specifies the style sheet language(Internet media type).");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vtype;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("style");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vtype = testNode.type;

      assert_equals(vtype,"text/css");
       
    domtest.done();
}));

</script>
<p>The type attribute specifies the style sheet language(Internet media type).

    Retrieve the type attribute and examine its value.</p>
<div id="log"></div>
</body></html>
